backend/requirement.txt
Flask==2.2.5
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.2.3
pymongo==4.4.0
